+++ src/components/LoginModal/LoginModal.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Box, Text } from '@chakra-ui/react';
 import { Button } from '@/components/Button'; 
 import { useRouter } from 'next/navigation';
